backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.7.0
google-cloud-aiplatform==1.38.1
google-cloud-firestore==2.13.1
google-cloud-speech==2.22.0
google-auth==2.23.4
google-api-python-client==2.108.0
firebase-admin==6.2.0
httpx==0.25.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
beautifulsoup4==4.12.2